fixed runtime errors

+++ application/src/main/java/com/amazon/proserve/application/flight/command/ChangeFlightNameCommand.java
@@ -5,13 +5,12 @@ import lombok.Value;
 import javax.validation.constraints.NotBlank;
 import javax.validation.constraints.NotNull;
 import javax.validation.constraints.PositiveOrZero;
-import java.math.BigInteger;
 
 @Value(staticConstructor = "of")
 public class ChangeFlightNameCommand {
     @NotNull
     @PositiveOrZero
-    BigInteger flightNo;
+    Long flightNo;
     @NotBlank
     String newFlightName;
 }
