+++ domain/src/main/java/com/amazon/proserve/domain/flight/vo/FlightNo.java
@@ -3,10 +3,9 @@ package com.amazon.proserve.domain.flight.vo;
 import lombok.Value;
 
 import javax.validation.constraints.NotNull;
-import java.math.BigInteger;
 
 @Value(staticConstructor = "of")
 public class FlightNo {
     @NotNull
-    BigInteger value;
+    Long value;
 }
