+++ infrastructure/src/main/java/com/amazon/proserve/infrastructure/flight/persistence/jpa/FlightJpaRepository.java
@@ -3,8 +3,6 @@ package com.amazon.proserve.infrastructure.flight.persistence.jpa;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import java.math.BigInteger;
-
 @Repository
-public interface FlightJpaRepository extends JpaRepository<FlightJpaEntity, BigInteger> {
+public interface FlightJpaRepository extends JpaRepository<FlightJpaEntity, Long> {
 }
