+++ interface/src/main/java/com/amazon/proserve/flight/api/dto/ChangeFlightNameRequest.java
@@ -3,8 +3,6 @@ package com.amazon.proserve.flight.api.dto;
 import lombok.*;
 import lombok.experimental.FieldDefaults;
 
-import java.math.BigInteger;
-
 @Builder
 @Getter
 @FieldDefaults(makeFinal = true, level = AccessLevel.PRIVATE)
@@ -12,6 +10,6 @@ import java.math.BigInteger;
 @ToString
 @EqualsAndHashCode
 public class ChangeFlightNameRequest {
-    private BigInteger flightNo;
+    private Long flightNo;
     private String flightName;
 }
